initial groundwork additions
