+++ package.json
@@ -14,6 +14,7 @@
     "@tailwindcss/vite": "^4.1.11",
     "@turf/boolean-valid": "^7.3.5",
     "@turf/helpers": "^7.3.5",
+    "@usace/groundwork": "^4.1.0",
     "d3-scale": "^4.0.2",
     "d3-scale-chromatic": "^3.1.0",
     "money-clip": "^3.0.5",
